Tsff 1499 barnetillegg hendelser brev (#365)
* test for barnetillegg hendelse
* tekster for barnetillegg
* barnetillegget fra UngdomsytelseSatser er sum for antall barn. Deler den derfor antall barn for å finne dagsatsen.

+++ formidling/src/main/java/no/nav/ung/sak/formidling/template/dto/innvilgelse/SatsEndringHendelseDto.java
@@ -8,7 +8,7 @@ public record SatsEndringHendelseDto(
     boolean dødsfallBarn,
     LocalDate fom,
     long dagsats,
-    int barnetillegg,
+    long barnetillegg,
     boolean fikkFlereBarn) {
 
 }
